refactor(desktop): 拆出 models 连接与删除对话框
将 ModelsSettingsPanel（含各连接/删除对话框）迁入 settings/models/models-settings-panel.tsx。
